requirements.txt
imdbpy>=2021.4.18
flake8>=4.0.1
jsonpickle>=2.0.0

